Converted project over to windows

+++ src/Rendering/RenderCommands/Commands/BlitRenderCommand.cpp
@@ -4,7 +4,7 @@
 
 #include "BlitRenderCommand.h"
 
-BlitRenderCommand::BlitRenderCommand(const RenderTarget &src, const RenderTarget& dest, unsigned long sortKey) noexcept
+BlitRenderCommand::BlitRenderCommand(const RenderTarget &src, const RenderTarget& dest, unsigned long long sortKey) noexcept
         : RenderCommand(Type::BLIT, sortKey), m_src(src), m_dest(dest)
 {
 }
