+++ src/Rendering/RenderCommands/Commands/SetRenderTargetRenderCommand.cpp
@@ -4,7 +4,7 @@
 
 #include "SetRenderTargetRenderCommand.h"
 
-SetRenderTargetRenderCommand::SetRenderTargetRenderCommand(RenderTarget rt, unsigned long sortKey)
+SetRenderTargetRenderCommand::SetRenderTargetRenderCommand(RenderTarget rt, unsigned long long sortKey)
 : RenderCommand(Type::BLIT, sortKey), m_rt(rt)
 {
